requirements.txt
groq==0.18.0
httpx==0.28.1
langchain==0.3.20
langchain_community==0.3.19
langchain_core==0.3.41
langchain_google_genai==2.0.11
langchain_groq==0.2.4
langchain_huggingface==0.1.2
langchain_text_splitters==0.3.6
langgraph==0.3.5
pandas==2.2.3
pydantic==2.10.6
PyYAML==6.0.2
streamlit==1.42.2
torch==2.6.0
transformers==4.48.2
typing_extensions==4.12.2
